requirements.txt
alpaca-trade-api>=3.0.0
pandas>=1.5.0
ta>=0.10.2
python-dotenv>=1.0.0
pytz>=2023.3

